refactor(structure): move ui-tokens from lib to config
ui-tokens.ts is pure static design-token class strings with no logic — it
belongs with the other static UI config (config/ui.ts z-index layers), not
under lib/ which is for functions/logic. Mechanical import path update across
19 consumers.
