+++ src/app/[locale]/lenses/[mount]/(browse)/collections/page.tsx
@@ -19,7 +19,7 @@ import {
 import { getLensesByMount } from "@/lib/lens-data";
 import { urlSegmentToMount, mountHasCollections } from "@/lib/mount";
 import { buildAlternates, defaultOgImages } from "@/lib/seo";
-import { ACTION_PRIMARY_CLS } from "@/lib/ui-tokens";
+import { ACTION_PRIMARY_CLS } from "@/config/ui-tokens";
 import LensIndexShell from "@/components/LensIndexShell";
 import CollectionChipRail from "@/components/CollectionChipRail";
